Add a script to extract RGB and Depth images from OpenSfM results

+++ scripts/utils.py
@@ -1,7 +1,7 @@
 import shutil
 from pathlib import PosixPath
 
-def make_output_dir(output_image_dir_path: PosixPath):
-    if output_image_dir_path.exists():
+def make_output_dir(output_image_dir_path: PosixPath, clean=False):
+    if output_image_dir_path.exists() and clean:
         shutil.rmtree(str(output_image_dir_path))
     output_image_dir_path.mkdir()
